+++ CHANGELOG.md
@@ -4,7 +4,11 @@
 - Rewrote the entire Database and Database export/import System.
 
 ## Features
+- added: Temporary Collections
+- added: Refresh Storage Info Button
+- improved: Better keyboard navbar navigation
 - improved: Database imports without deleting whole Database
+- improved: Manual Download text
 - removed: Empty collection without deleting
 - removed: all previous user data without warning